requirements.txt
Django~=4.0.4
asgiref==3.5.0
djangorestframework==3.13.1
psycopg2-binary==2.9.3
pytz==2022.1
sqlparse == 0.4.2
tzdata == 2022.1
